requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
PyMuPDF
scikit-learn
pytesseract
Pillow
